requirements.txt
pytrends==4.9.2
pandas>=2.0.0
plotly>=5.17.0
jinja2>=3.1.2
tenacity>=8.2.3
playwright>=1.40.0
requests>=2.31.0
pyyaml>=6.0
